--- src/subscriptions/types.ts
@@ -1,14 +1,0 @@
-export interface SubscriptionMatchMessage {
-    subscriptionId: string;
-    tenantId?: string;
-    channelType: string;
-    endpoint: string;
-    channelPayload?: string;
-    channelHeader: string[];
-    matchedResource: {
-        id: string;
-        resourceType: string;
-        versionId: string;
-        lastUpdated: string;
-    };
-}
