--- subscriptions/index.ts
@@ -1,3 +1,0 @@
-import { sendRestHookNotification } from './restHook';
-
-exports.handler = sendRestHookNotification;
